+++ server/templates/crowdin.js
@@ -1,3 +1,3 @@
 /* Crowdin configuration */
-const _jipt = [];
+const _jipt = []; // eslint-disable-line no-underscore-dangle
 _jipt.push(['project', 'BRANDING']);