+++ server/templates/matomo.js
@@ -1,6 +1,6 @@
 var _paq = window._paq || [];
 /* tracker methods like "setCustomDimension" should be called before "trackPageView" */
-_paq.push(["setDoNotTrack", true]);
+_paq.push(['setDoNotTrack', true]);
 _paq.push(['requireConsent']);
 _paq.push(['trackPageView']);
 (function () {