+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
+## Unreleased
+### Engineering Use
+- YLP-107 Create lambda edge generatation script for CloudFront
+
 ## 1.3.1 2020-08-19
 ### Engineering
 - PT-1470 Update blip docker image to run as Coreye user
@@ -17,7 +21,7 @@ It is based on Tidepool Blip 1.27.
 ## 1.2.0 - 2020-07-31
 ### Changed
 - YLP-72 Have two distincts fields for first name and last name in profile page
-### Engineering
+### Engineering Use
 - YLP-48 Change crowdin pseudo language from 'it' to 'lol'
 - Externalize languages definition (out of code)